refactor state to App.tsx, then passing prop to child page
